requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg[binary]>=3.1.18
pydantic-settings>=2.2.1
typer>=0.12.3
websockets>=12.0
pandas>=2.2.2
openpyxl>=3.1.2
httpx>=0.27,<0.28
python-multipart>=0.0.9
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
python-magic>=0.4.27; platform_system != 'Windows'
python-magic-bin>=0.4.14; platform_system == 'Windows'
slowapi>=0.1.9
rapidfuzz>=3.9.0
aiosqlite>=0.20.0

# extras: dev
ruff>=0.4.2
black>=24.3.0
pytest>=8.1.0
pytest-asyncio>=0.23
python-dotenv>=1.0.0
Pillow>=10.4.0
beautifulsoup4>=4.12.3
clamd>=1.0.2
dramatiq[redis]>=1.15.0
rembg>=2.0.57
tenacity>=8.2.3
playwright>=1.42.0
weasyprint>=61.0

# PDF/OCR processing
pdfplumber>=0.11  # fija versión reciente que usa pypdf moderno
pypdf>=4.3  # explícito: pypdf reciente evita warnings ARC4 con cryptography
camelot-py[cv]
ocrmypdf
pdf2image
pytesseract
opencv-python-headless
onnxruntime>=1.22.1  # runtime para rembg (model inference)

# Compatibilidad Camelot: requiere API antigua de PyPDF2 (PdfFileReader); fijamos <3
PyPDF2>=2.10,<3

# LLM proveedor OpenAI
openai>=1.40.0
notion-client>=2.2.1
